Move trusted_hosts logic to PipSession.
