+++ src/pip/_internal/commands/list.py
@@ -126,7 +126,6 @@ class ListCommand(IndexGroupCommand):
         return PackageFinder.create(
             search_scope=search_scope,
             selection_prefs=selection_prefs,
-            trusted_hosts=options.trusted_hosts,
             session=session,
         )
