+++ src/pip/_internal/req/req_file.py
@@ -272,7 +272,7 @@ def process_line(
             finder.set_allow_all_prereleases()
         for host in opts.trusted_hosts or []:
             source = 'line {} of {}'.format(line_number, filename)
-            finder.add_trusted_host(host, source=source)
+            session.add_trusted_host(host, source=source)
 
 
 def break_args_options(line):
